# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Yukakad API
  version: 1.0.0
  description: API SaaS undangan digital Yukakad.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /invitations/{id}/domains:
    parameters: [{ $ref: '#/components/parameters/InvitationID' }]
    get: { responses: { '200': { description: Custom domain list } } }
    post: { responses: { '201': { description: Custom domain created } } }
  /invitations/{id}/domains/{domain_id}/verify:
    parameters:
      - { $ref: '#/components/parameters/InvitationID' }
      - name: domain_id
        in: path
        required: true
        schema: { type: string }
    post: { responses: { '200': { description: Domain verified } } }
  /templates:
    get:
      security: []
      parameters:
        - { name: event_type, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: tier, in: query, schema: { type: string, enum: [free, premium, signature] } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { '200': { description: Filtered template catalog } }
  /templates/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      security: []
      responses: { '200': { description: Template detail and preview metadata } }
...
  /admin/templates:
    get:
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [draft, active, archived] } }
        - { name: event_type, in: query, schema: { type: string } }
        - { name: tier, in: query, schema: { type: string, enum: [free, premium, signature] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { '200': { description: Admin template catalog } }
    post: { responses: { '201': { description: Template created } } }
  /admin/templates/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { responses: { '200': { description: Template detail } } }
    patch: { responses: { '200': { description: Template updated } } }
    delete: { responses: { '204': { description: Template deleted } } }
  /admin/templates/{id}/preview:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      requestBody:
        required: true
        content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } }
      responses: { '200': { description: Preview uploaded } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    InvitationID: { name: id, in: path, required: true, schema: { type: string } }
...